+++ src/main/groovy/org/codenarc/rule/unnecessary/UnnecessaryReturnKeywordRule.groovy
@@ -60,8 +60,7 @@ class UnnecessaryReturnKeywordAstVisitor extends AbstractAstVisitor {
     /**
      * This is not a good general function for AstUtils.
      * It is too specific and may not work across different ASTNode subtypes.
-     * @param node
-     *      node
+     * @param node the ASTNode whose last statement we need
      */
     private static Statement getLastStatement(ASTNode node) {
         if (node.code instanceof BlockStatement && node.code.statements) {